% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{adjustProteinLevel}
\alias{adjustProteinLevel}
\title{Adjust differential analysis result with respect to protein abundance}
\usage{
adjustProteinLevel(diffSite, diffProtein)
}
\arguments{
\item{diffSite}{A data frame for the differential analysis result of PTMs,
returned by the function \code{extractMeanDiff} with the option
\code{per_protein = FALSE}. The data frame contains columns of
\code{Protein}, \code{Site}, \code{Label}, \code{log2FC}, \code{SE},
\code{Tvalue}, \code{DF}, and \code{pvalue}.}

\item{diffProtein}{A data frame for the differential analysis result of
proteins, returned by the function \code{extractMeanDiff} with the option
\code{per_protein = TRUE}. The data frame contains columns of
\code{Protein}, \code{Label}, \code{log2FC}, \code{SE}, \code{Tvalue},
\code{DF}, and \code{pvalue}.}
}
\value{
A data frame.
}
\description{
\code{adjustProteinLevel} performs the adjustment with respect to protein
abundance.
}
