% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{annot_site}
\alias{annot_site}
\title{Annotate modification site}
\usage{
annot_site(aa_idx, residue, len_idx = NULL)
}
\arguments{
\item{aa_idx}{An integer vector. Location of the sites.}

\item{residue}{A string vector. Amino acid residue.}

\item{len_idx}{An integer. Default is \code{NULL}}
}
\value{
A string.
}
\description{
\code{annot_site} annotates modified sites as their residues and locations.
}
\examples{
annot_site(10, "K")
annot_site(10, "K", 3L)
}
