% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatePTM.R
\name{annotSite}
\alias{annotSite}
\title{Annotate modification site}
\usage{
annotSite(aaIndex, residue, lenIndex = NULL)
}
\arguments{
\item{aaIndex}{An integer vector. Location of the sites.}

\item{residue}{A string vector. Amino acid residue.}

\item{lenIndex}{An integer. Default is \code{NULL}}
}
\value{
A string.
}
\description{
\code{annotSite} annotates modified sites as their residues and locations.
}
\examples{
annotSite(10, "K")
annotSite(10, "K", 3L)

}
