% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mod.R
\name{compare_mod}
\alias{compare_mod}
\title{Compare abundance of a modified site across conditions.}
\usage{
compare_mod(df_mod, controls, cases, protadj = TRUE)
}
\arguments{
\item{df_mod}{A data frame.}

\item{controls}{Control group as reference.}

\item{cases}{Case group.}

\item{protadj}{A logical. \code{TRUE} performs protein-level adjustment,
\code{FALSE} otherwise. Default is \code{TRUE}.}
}
\value{
A data frame.
}
\description{
\code{compare_mod} performs significance analysis for differentially modified
sites across conditions.
}
\examples{
compare_mod(df_mod, controls, cases, protadj = TRUE)
}
