% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mod2.R
\name{compare_mod2}
\alias{compare_mod2}
\title{Compare abundance of a modified site across conditions.}
\usage{
compare_mod2(df_mod, grp_ctrl, grp_case, protadj = TRUE)
}
\arguments{
\item{df_mod}{A data frame.}

\item{grp_ctrl}{Control group as reference.}

\item{grp_case}{Case group.}

\item{protadj}{A logical. \code{TRUE} performs protein-level adjustment,
\code{FALSE} otherwise. Default is \code{TRUE}.}
}
\value{
A data frame.
}
\description{
\code{compare_mod} performs significance analysis for differentially modified
sites across conditions.
}
\examples{
compare_mod(df_mod, grp_ctrl, grp_case, protadj = TRUE)

}
