% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{.aggregateFC}
\alias{.aggregateFC}
\title{Aggregate FC estimates from multiple sources by averaging over them}
\usage{
.aggregateFC(data)
}
\description{
Aggregate FC estimates from multiple sources by averaging over them
}
\keyword{internal}
