% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designSampleSizePTM.R
\name{.calculatePowerPTM}
\alias{.calculatePowerPTM}
\title{Power calculation for PTM experiment}
\usage{
.calculatePowerPTM(
  desiredFC,
  FDR,
  delta,
  ptm_median_sigma_error,
  protein_median_sigma_error,
  ptm_median_sigma_subject,
  protein_median_sigma_subject,
  numSample
)
}
\arguments{
\item{desiredFC}{the range of a desired fold change which includes the lower
and upper values of the desired fold change.}

\item{FDR}{a pre-specified false discovery ratio (FDR) to control the overall
false positive rate. Default is 0.05}

\item{numSample}{minimal number of biological replicates per condition.
TRUE represents you require to calculate the sample size for this category,
else you should input the exact number of biological replicates.}
}
\value{
\code{float} of power
}
\description{
Power calculation for PTM experiment
}
\keyword{internal}
