% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_converters.R
\name{.fixTerminus}
\alias{.fixTerminus}
\title{Add site location and aa}
\usage{
.fixTerminus(data, terminus_id, unmod_pep_col)
}
\arguments{
\item{data}{data.table}

\item{fasta_file}{string or data.table}
}
\value{
data.table
}
\description{
Add site location and aa
}
\keyword{internal}
