% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_converters.R
\name{.locateSites}
\alias{.locateSites}
\title{Add site location and aa}
\usage{
.locateSites(
  data,
  mod_id,
  protein_name_col,
  unmod_pep_col,
  mod_pep_col,
  mod_id_is_numeric
)
}
\arguments{
\item{data}{data.table}

\item{mod_id}{string}
}
\value{
data.table
}
\description{
Add site location and aa
}
\keyword{internal}
