% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_param.R
\name{extract_param}
\alias{extract_param}
\title{Extract estimated parameters.}
\usage{
extract_param(nested)
}
\arguments{
\item{nested}{An instance of nested data frame.}
}
\value{
A data frame restoring the estimated model parameters.
}
\description{
\code{extract_param} extracts the estimated model parameters from nested data
frame.
}
\examples{
extract_param(nested)

}
