% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_censored_aft.R
\name{fill_censored_aft}
\alias{fill_censored_aft}
\title{Annotate censored values and fill with AFT.}
\usage{
fill_censored_aft(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame.
}
\description{
\code{fill_censored_aft} annotates censored feature intensities and imputes
the censored values using the accelerated failure time model.
}
\examples{
fill_censored_aft(data)
}
