% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{fixedGroup}
\alias{fixedGroup}
\title{Linear model with group effect}
\usage{
fixedGroup(df)
}
\arguments{
\item{df}{A data frame with columns \code{log2inty} and \code{group} for one
PTM site.}
}
\value{
An \code{lm} model object.
}
\description{
\code{fixedGroup} fits and returns a linear model with \code{group} as a
fixed effect.
}
\examples{
x <- data.frame(
    group=rep(c("1", "2"), 3),
    log2inty=rep(c(10, 12), 3) + rnorm(6)
)
fixedGroup(x)

}
