% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{fixedGroupBatch}
\alias{fixedGroupBatch}
\title{Linear model with group and batch effects}
\usage{
fixedGroupBatch(df)
}
\arguments{
\item{df}{A data frame with columns \code{log2inty}, \code{group}, and
\code{batch} for one PTM site.}
}
\value{
An \code{lm} model object.
}
\description{
\code{fixedGroupBatch} fits and returns a linear model with \code{log2inty}
as response, and \code{group} and \code{batch} as fixed effects.
}
\examples{
x <- data.frame(
    batch=rep(c("1", "2"), each=4),
    group=rep(c("1", "2"), 4),
    log2inty=rep(c(10, 12), 4) + rnorm(8)
)
fixedGroupBatch(x)

}
