% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{fragpipe_input}
\alias{fragpipe_input}
\title{Output of FragPipe TMT PTM experiment}
\format{
A data.table with 29 columns and 246 rows.
}
\usage{
fragpipe_input
}
\description{
This dataset was provided by the FragPipe team at the Nesvilab. It was
processed using Philosopher and targeted Phosphorylation.
}
\examples{
head(fragpipe_input)
}
\keyword{datasets}
