% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{fragpipe_input_protein}
\alias{fragpipe_input_protein}
\title{Output of FragPipe TMT global profiling experiment}
\format{
A data.table with 27 columns and 47 rows.
}
\usage{
fragpipe_input_protein
}
\description{
This dataset was provided by the FragPipe team at the Nesvilab. It was
processed using Philosopher and targeted Phosphorylation.
}
\examples{
head(fragpipe_input_protein)
}
\keyword{datasets}
