% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_site.R
\name{locate_site}
\alias{locate_site}
\title{Locate potential modification sites.}
\usage{
locate_site(pep_seq, pep_idx, mod_res)
}
\arguments{
\item{pep_seq}{A string. Peptide sequence.}

\item{pep_idx}{A matrix. Starting and ending indices of the peptide.}

\item{mod_res}{A string. Modification residue.}
}
\value{
A string.
}
\description{
\code{locate_site} locates modifiable sites on a peptide.
}
\examples{
locate_site(pep_seq, pep_idx, mod_res)

}
