% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatePTM.R
\name{locateMod}
\alias{locateMod}
\title{Locate modified sites with a peptide}
\usage{
locateMod(peptide, aaStart, residueSymbol)
}
\arguments{
\item{peptide}{A string. Peptide sequence.}

\item{aaStart}{An integer. Starting index of the peptide.}

\item{residueSymbol}{A string. Modification residue and denoted symbol.}
}
\value{
A string.
}
\description{
\code{locateMod} locates modified sites with a peptide.
}
\examples{
locateMod("P*EP*TIDE", 3, "\\\\*")

}
