% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{maxq_tmt_annotation}
\alias{maxq_tmt_annotation}
\title{Example annotation file for a TMT MaxQuant experiment.}
\format{
A data.table with 7 columns.
}
\usage{
maxq_tmt_annotation
}
\description{
Must be manually created by the user and input into the
MaxQtoMSstatsPTMFormat converter. Requires the correct columns and maps the
experimental desing into the MSstats format. Specify unique bioreplicates for
group comparison designs, and the same bioreplicate for repeated measure
designs. The columns and descriptions are below.
}
\details{
\itemize{
\item Run : Run name that matches exactly with MaxQuant run. Used to join
evidence and metadata in annotation file.
\item Fraction : If multiple fractions were used (i.e. the same mixture
split into multiple fractions) enter that here.
TechRepMixture : Multiple runs using the same bioreplicate
\item Channel : Mixture channel used
\item Condition : Name of condition that was used for each run.
\item Mixture : The unique mixture (plex) name
\item BioReplicate : Name of biological replicate. Repeating the same name
here will tell MSstatsPTM that the experiment is a repeated measure design.
}
}
\examples{
head(maxq_tmt_annotation)
}
\keyword{datasets}
