% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{maxq_tmt_evidence}
\alias{maxq_tmt_evidence}
\title{Example MaxQuant evidence file from the output of a TMT experiment}
\format{
a data.table with 96 columns and 199 rows, the output of MaxQuant
}
\usage{
maxq_tmt_evidence
}
\description{
Experiment was performed by the Olsen lab and published on Nat. Commun.
(citation below).
}
\details{
Hogrebe, A., von Stechow, L., Bekker-Jensen, D.B. et al. Benchmarking common
quantification strategies for large-scale phosphoproteomics. Nat Commun 9,
1045 (2018). https://doi.org/10.1038/s41467-018-03309-6

The experiment was processed using MaxQuant by the computational proteomics
team at Pfizer (Liang Xue and Pierre Jean).

The experiment did not contain a global profiling run, but we show an example
of extracting the unmodified peptides and using them in place of the
profiling run.
}
\examples{
head(maxq_tmt_evidence)
}
\keyword{datasets}
