% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQtoMSstatsPTMFormat.R
\name{MaxQtoMSstatsPTMFormat}
\alias{MaxQtoMSstatsPTMFormat}
\title{Convert output of TMT labeled MaxQuant experiment into MSstatsPTM format}
\usage{
MaxQtoMSstatsPTMFormat(
  sites.data,
  annotation,
  evidence = NULL,
  proteinGroups = NULL,
  mod.num = "Single",
  keyword = "phos",
  which.proteinid.ptm = "Protein",
  which.proteinid.protein = "Leading.razor.protein",
  removeMpeptides = FALSE
)
}
\arguments{
\item{sites.data}{modified peptide output from MaxQuant. For example, a
phosphorylation experiment would require the Phospho(STY)Sites.txt file}

\item{annotation}{data frame which contains column Run, Fraction,
TechRepMixture, Mixture, Channel, BioReplicate, Condition.}

\item{evidence}{for global protein dataset. name of 'evidence.txt' data,
which includes feature-level data.}

\item{proteinGroups}{for global protein dataset, name of 'proteinGroups.txt'
data.}

\item{mod.num}{For modified peptide dataset. The number modifications per
peptide to be used. If "Single", only peptides with one modification will be
used. Otherwise "Total" can be selected which does not cap the number of
modifications per peptide. "Single" is the default. Selecting "Total" may
confound the effect of different modifications.}

\item{keyword}{the sub-name of columns in the sites.data file. For
phosphorylation data, this value should be "phos". The default is "phos".}

\item{which.proteinid.ptm}{For PTM dataset, which column to use for protein
name. Use 'Proteins'(default) column for protein name. 'Leading.proteins' or
'Leading.razor.protein' or 'Gene.names' can be used instead to get the
protein ID with single protein. However, those can potentially have the
shared peptides.}

\item{which.proteinid.protein}{For Protein dataset, which column to use for
protein name. Same options as above.}

\item{removeMpeptides}{If Oxidation (M) modifications should be removed.
Default is TRUE.}
}
\value{
a list of two data.tables named 'PTM' and 'PROTEIN' in the format
required by MSstatsPTM.
}
\description{
Takes as input TMT experiments from MaxQ and converts the data into the
format needed for MSstatsPTM. Requires only the modified file from MaxQ (for
example Phospho(STY)Sites) and an annotation file for PTM data. To adjust
modified peptides for changes in global protein level, unmodified TMT
experimental data must also be returned.
}
\examples{

head(raw.input.tmt$PTM)
head(raw.input.tmt$PROTEIN)

}
