% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ptm.R
\name{model_ptm}
\alias{model_ptm}
\title{Whole-plot modeling for all sites.}
\usage{
model_ptm(df_sum, w_batch = FALSE)
}
\arguments{
\item{df_sum}{A data frame.}

\item{w_batch}{A logical. \code{TRUE} considers batch effect, \code{FALSE}
otherwise. Default is \code{FALSE}.}
}
\value{
A data frame.
}
\description{
\code{model_ptm} fits a whole-plot model for all sites in consideration of
the experimental design and the underlying protein abundance, and returns
the estimates of model parameters.
}
\examples{
model_ptm(df_sum, w_batch = FALSE)
}
