% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_site.R
\name{nest_site}
\alias{nest_site}
\title{Whole-plot modeling.}
\usage{
nest_site(df_sum, w_batch = FALSE)
}
\arguments{
\item{df_sum}{A data frame.}

\item{w_batch}{A logical. \code{TRUE} considers batch effect, \code{FALSE}
otherwise. Default is \code{FALSE}.}
}
\value{
An instance of nested data frame.
}
\description{
\code{nest_site} fits and returns a whole-plot model for each site in
consideration of the experimental design, organized in a nested data frame.
}
\examples{
nest_site(df_sum, w_batch = FALSE)

}
