% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_ptm.R
\name{normalize_ptm}
\alias{normalize_ptm}
\title{Normalize feature intensities across runs.}
\usage{
normalize_ptm(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame.
}
\description{
\code{normalize_ptm} normalizes feature intensities across runs in a PTM
dataset, using unpaired unmodified peptides.
}
\examples{
normalize_ptm(data)
}
