% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pd_annotation}
\alias{pd_annotation}
\title{Example annotation file for a label-free Proteome Discoverer experiment.}
\format{
A data.table with 3 columns.
}
\usage{
pd_annotation
}
\description{
Must be manually created by the user and input into the
PDtoMSstatsPTMFormat converter. Requires the correct columns and maps the
experimental desing into the MSstats format. Specify unique bioreplicates for
group comparison designs, and the same bioreplicate for repeated measure
designs. The columns and descriptions are below.
}
\details{
\itemize{
\item Run : Run name that matches exactly with PD run. Used to join
evidence and metadata in annotation file.
\item Condition : Name of condition that was used for each run.
\item BioReplicate : Name of biological replicate. Repeating the same name
here will tell MSstatsPTM that the experiment is a repeated measure design.
}
}
\examples{
head(pd_annotation)
}
\keyword{datasets}
