% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pd_psm_input}
\alias{pd_psm_input}
\title{Example Proteome Discoverer evidence file from the output of a label free experiment}
\format{
a data.table with 60 columns and 1657 rows, the output of PD
}
\usage{
pd_psm_input
}
\description{
Experiment was performed by the Olsen lab and published on Nat. Commun.
(citation below).
}
\details{
Bekker-Jensen, D.B., Bernhardt, O.M., Hogrebe, A. et al. Rapid and
site-specific deep phosphoproteome profiling by data-independent acquisition
without the need for spectral libraries. Nat Commun 11, 787 (2020).
https://doi.org/10.1038/s41467-020-14609-1

The experiment was processed using Proteome Discoverer by the computational
proteomics team at Pfizer (Liang Xue and Pierre Jean).

The experiment did not contain a global profiling run, but we show an example
of extracting the unmodified peptides and using them in place of the
profiling run.
}
\examples{
head(pd_psm_input)
}
\keyword{datasets}
