% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pd_testing_output}
\alias{pd_testing_output}
\title{Example output of Proteome Discoverer converter}
\format{
a list with 2 data.frames
}
\usage{
pd_testing_output
}
\description{
output using example data provided in package
}
\details{
The experiment did not contain a global profiling run, but we show an example
of extracting the unmodified peptides and using them in place of the
profiling run.
}
\examples{
head(pd_testing_output)
}
\keyword{datasets}
