% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTMestimate.R
\name{PTMestimateAbundance}
\alias{PTMestimateAbundance}
\title{Estimate log2-abundances of PTM sites.}
\usage{
PTMestimateAbundance(df, fac_batch = FALSE)
}
\arguments{
\item{df}{A data frame with columns of \code{protein}, \code{site},
\code{group}, \code{run}, \code{log2inty}, and possibly, \code{batch}.}

\item{fac_batch}{A logical. \code{TRUE} considers a fixed batch effect,
\code{FALSE} otherwise. Default is \code{FALSE}.}
}
\value{
A list.
}
\description{
\code{PTMestimateAbundance} takes as input the summarized log2-intensities
for each PTM site, performs statistical modeling for the abundance of the
site, and returns the estimates of model parameters for all sites in all
experimental conditions.
}
\examples{
PTMestimateAbundance(df, fac_batch = FALSE)
}
