% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{PTMlocate}
\alias{PTMlocate}
\title{Annotate modified sites with associated peptides}
\usage{
PTMlocate(peptide, uniprot, fasta, mod_residue, mod_symbol, rm_cnfnd = FALSE)
}
\arguments{
\item{peptide}{A string vector of peptide sequences. The peptide sequence
does not include its preceding and following AAs.}

\item{uniprot}{A string vector of Uniprot identifiers of the peptides'
originating proteins. UniProtKB entry isoform sequence is used.}

\item{fasta}{A tibble with FASTA information. Output of \code{tidy_fasta}.}

\item{mod_residue}{A string. Modifiable amino acid residues.}

\item{mod_symbol}{A string. Symbol of a modified site.}

\item{rm_cnfnd}{A logical. \code{TRUE} removes confounded
unmodified sites, \code{FALSE} otherwise. Default is \code{FALSE}.}
}
\value{
A data frame with three columns: \code{uniprot_iso}, \code{peptide},
\code{site}.
}
\description{
\code{PTMlocate} annotates modified sites with associated peptides.
}
