% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{PTMnormalize}
\alias{PTMnormalize}
\title{Normalization of log2-intensities across MS runs}
\usage{
PTMnormalize(df, method = "median", calc = "all", reference)
}
\arguments{
\item{df}{A data frame contains all the following columns: \code{run},
\code{feature}, \code{is_mod}, and \code{log2inty}.}

\item{method}{A string defining the normalization method. Default is
\code{"median"}, which equalizes the medians of log2-intensities across MS
runs. Other methods include to equalize log2 of intensity summation
(\code{"logsum"}), to equalize the means of log2-intensities
(\code{"mean"}), and to adjust the log2-intensities based on a reference
(\code{"ref"}) given by (\code{reference}).}

\item{calc}{A string defining parts of log2-intensities used for calculating
the adjustment value for each MS run. Default is \code{"all"}, which uses
all measurements. Other options include to use only log2-intensities from
peptides with PTMs of interest (\code{"mod"}) and log2-intensities from
unmodified peptides (\code{"unmod"}).}

\item{reference}{A data frame defining the adjustment of log2-intensities for
each MS runs, with columns of \code{run} and \code{adjLog2inty}.}
}
\value{
A data frame with same columns as in \code{df}.
}
\description{
\code{PTMnormalize} normalizes log2-intensities of spectral features across
MS runs.
}
\examples{
\dontrun{
PTMnormalize(df)
}
}
