% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{PTMsimulateExperiment}
\alias{PTMsimulateExperiment}
\title{Simulate PTM quantification experiments}
\usage{
PTMsimulateExperiment(nGroup, nRep, nProtein, nSite, nFeature, logAbundance)
}
\arguments{
\item{nGroup}{An integer to specify the number of conditions.}

\item{nRep}{An integer to specify the number of replicates per condition.}

\item{nProtein}{An integer to specify the number of protein.}

\item{nSite}{An integer to specify the number of PTM sites per protein.}

\item{nFeature}{An integer to specify the number of features per site.}

\item{logAbundance}{A list of two lists named \code{PTM} and \code{PROTEIN}.
Each contains four elements: \code{mu} (a numeric representing the overall
mean log2-abundance), \code{delta} (a numeric vector for the deviation of
the mean log2-abundance for each group from the overall mean), \code{sRep}
(a numeric representing the standard deviation for run-to-run variation),
and \code{sPeak} (a numeric representing the standard deviation in peak
log2-intensities).}
}
\value{
A tibble with columns of \code{protein}, \code{site}, \code{group},
\code{run}, \code{feature}, \code{log2inty}.
}
\description{
\code{PTMsimulateExperiment} simulates a PTM quantification experiment with
a list of log2-intensities of multiple spectral features, PTM sites and the
corresponding proteins, in multiple MS runs and conditions.
}
\examples{
PTMsimulateExperiment(
    nGroup=2, nRep=2, nProtein=1, nSite=1, nFeature=5,
    logAbundance=list(
        PTM=list(mu=25, delta=c(0, 1), sRep=0.2, sPeak=0.05),
        PROTEIN=list(mu=25, delta=c(0, 1), sRep=0.2, sPeak=0.05)
    )
)

}
