% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{PTMsummarize}
\alias{PTMsummarize}
\title{Site-level summarization}
\usage{
PTMsummarize(df, method = "tmp")
}
\arguments{
\item{df}{A data frame with columns of \code{protein}, \code{site},
\code{group}, \code{run}, \code{feature}, \code{log2inty}, and possibly,
\code{batch}.}

\item{method}{A string defining the summarization method. Default is
\code{"tmp"}, which applies Tukey's median polish. Other methods include
log2 of intensity summation (\code{"logsum"}), and mean (\code{"mean"}),
median (\code{"median"}) and max (\code{"max"}) of the log-intensities.}
}
\value{
A data frame with columns of \code{protein}, \code{site},
\code{group}, \code{run}, \code{log2inty}, and possibly, \code{batch}.
}
\description{
\code{PTMsummarize} summarizes log2-intensities of spectral features for each
PTM site into one value per run.
}
\examples{
\dontrun{
PTMsummarize(df)
}
}
