% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{raw.input}
\alias{raw.input}
\title{Example of input PTM dataset for LabelFree/DDA/DIA experiments.}
\format{
A list of two data.tables named PTM and PROTEIN with 1745 and 478
rows respectively.
}
\usage{
raw.input
}
\description{
It can be the output of MSstatsPTM converter ProgenesistoMSstatsPTMFormat or
other MSstats converter functions (Please see MSstatsPTM_LabelFree_Workflow
vignette). The dataset is formatted as a list with two data.tables named PTM
and PROTEIN. In each data.table the variables are as follows:
}
\details{
\itemize{
#'   \item ProteinName : Name of protein with modification site mapped in with
an underscore. ie "Protein_4_Y474"
\item PeptideSequence
\item Condition : Condition (ex. Healthy, Cancer, Time0)
\item BioReplicate : Unique ID for biological subject.
\item Run : MS run ID.
\item Intensity
\item PrecursorCharge
\item FragmentIon
\item ProductCharge
\item IsotopeLabelType
}
}
\examples{
head(raw.input$PTM)
head(raw.input$PROTEIN)

}
\keyword{datasets}
