% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulatePeaks}
\alias{simulatePeaks}
\title{Simulate peak log2-intensities}
\usage{
simulatePeaks(nGroup, nRep, nFeature, mu, delta, sRep, sPeak)
}
\arguments{
\item{nGroup}{An integer to specify the number of conditions.}

\item{nRep}{An integer to specify the number of replicates per condition.}

\item{nFeature}{An integer to specify the number of features per site.}

\item{mu}{A numeric to specify the overall mean log2-intensity.}

\item{delta}{A numeric to specify the deviation of the mean log2-abundance of
each group from the overall mean.}

\item{sRep}{A numeric to specify the standard deviation for run-to-run
variation.}

\item{sPeak}{A numeric to specify the standard deviation in peak
log2-intensities.}
}
\value{
A tibble with columns of \code{group}, \code{run}, \code{feature},
and \code{log2inty}.
}
\description{
\code{simulateSites} simulates a list of log2-intensities of multiple
spectral features of a PTM site, in multiple MS runs and conditions.
}
\examples{
simulatePeaks(nGroup = 2, nRep = 3, nFeature = 5, 25, c(0, 1), 0.2, 0.05)

}
