% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateSummarization}
\alias{simulateSummarization}
\title{Simulate site-level summarization for PTM experiment}
\usage{
simulateSummarization(nGroup, nRep, mu, delta, sRep)
}
\arguments{
\item{nGroup}{An integer to specify the number of conditions.}

\item{nRep}{An integer to specify the number of replicates per condition.}

\item{mu}{A numeric value of the overall mean log2-abundance.}

\item{delta}{A numeric vector to specify the deviation of the mean
log2-abundance of each group from the overall mean.}

\item{sRep}{A numeric. Standard deviation of the log2-intensities.}
}
\value{
A tibble with columns of \code{group}, \code{run} and \code{log2inty}.
}
\description{
\code{simulateSummarization} simulates the summarized log2-intensity value
of a PTM site in each MS run. The value is randomly generated based on a
normal distribtuion, where the average log2-intensity is defined for each
condition
}
\examples{
simulateSummarization(nGroup = 2, nRep = 3, 25, c(0, 1), 0.2)
}
