% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{spectronaut_annotation}
\alias{spectronaut_annotation}
\title{Example annotation file for a label-free Spectronaut experiment.}
\format{
A data.table with 5 columns.
}
\usage{
spectronaut_annotation
}
\description{
Must be manually created by the user and input into the
SpectronauttoMSstatsPTMFormat converter. Requires the correct columns and
maps the experimental desing into the MSstats format. Specify unique
bioreplicates for group comparison designs, and the same bioreplicate for
repeated measure designs. The columns and descriptions are below.
}
\details{
\itemize{
\item Run : Run name that matches exactly with Spectronaut run. Used to join
evidence and metadata in annotation file.
\item Condition : Name of condition that was used for each run.
\item BioReplicate : Name of biological replicate. Repeating the same name
here will tell MSstatsPTM that the experiment is a repeated measure design.
\item Raw.file : Run name that matches exactly with Spectronaut run. Used to join
evidence and metadata in annotation file.
}
}
\examples{
head(spectronaut_annotation)
}
\keyword{datasets}
