% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{spectronaut_input}
\alias{spectronaut_input}
\title{Example Spectronaut evidence file from the output of a label free experiment}
\format{
a data.table with 23 columns and 2683 rows, the output of Spectronaut
}
\usage{
spectronaut_input
}
\description{
Experiment was performed by the Olsen lab and published on Nat. Commun.
(citation below).
}
\details{
Bekker-Jensen, D.B., Bernhardt, O.M., Hogrebe, A. et al. Rapid and
site-specific deep phosphoproteome profiling by data-independent acquisition
without the need for spectral libraries. Nat Commun 11, 787 (2020).
https://doi.org/10.1038/s41467-020-14609-1

The experiment was processed using Spectronaut by the computational proteomics
team at Pfizer (Liang Xue and Pierre Jean).

The experiment did not contain a global profiling run, but we show an example
of extracting the unmodified peptides and using them in place of the
profiling run.
}
\examples{
head(spectronaut_input)
}
\keyword{datasets}
