% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_bch.R
\name{tidy_bch}
\alias{tidy_bch}
\title{Extract estimate of group effect.}
\usage{
tidy_bch(bch_fit, df_bch)
}
\arguments{
\item{bch_fit}{An object of class \code{lm}.}

\item{df_bch}{A number. Degrees of freedom.}
}
\value{
A data frame restoring the estimated model parameters.
}
\description{
\code{tidy_bch} extracts the estimate of group effect.
}
\examples{
tidy_bch(bch_fit, df_bch)

}
