% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_fasta.R
\name{tidy_fasta}
\alias{tidy_fasta}
\title{Read and tidy FASTA file.}
\usage{
tidy_fasta(path)
}
\arguments{
\item{path}{A string. Path to the FASTA file.}
}
\value{
A data frame.
}
\description{
\code{tidy_fasta} reads and tidys FASTA file.
}
\examples{
tidy_fasta(path)

}
