% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_fasta.R
\name{tidy_fasta2}
\alias{tidy_fasta2}
\title{Read and tidy a FASTA file.}
\usage{
tidy_fasta2(aa_file)
}
\arguments{
\item{aa_file}{A string of path to a FASTA file.}
}
\value{
A tibble with columns named \code{header}, \code{sequence},
\code{uniprot_ac}, \code{uniprot_iso}, \code{entry_name}.
}
\description{
\code{tidy_fasta} reads and tidys FASTA file.
}
\examples{
tidy_fasta(aa_file)
}
