% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{tidyFasta}
\alias{tidyFasta}
\title{Read and tidy a FASTA file}
\usage{
tidyFasta(path)
}
\arguments{
\item{path}{A string of path to a FASTA file.}
}
\value{
A tibble with columns named \code{header}, \code{sequence},
\code{uniprot_ac}, \code{uniprot_iso}, \code{entry_name}.
}
\description{
\code{tidyFasta} reads and tidys FASTA file.
}
\examples{
tidyFasta("https://www.uniprot.org/uniprot/O13297.fasta")

}
