% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{adjustProteinLevel}
\alias{adjustProteinLevel}
\title{Adjust differential analysis result with respect to protein abundance}
\usage{
adjustProteinLevel(diffSite, diffProtein)
}
\arguments{
\item{diffSite}{A data frame for the differential analysis result of PTMs.}

\item{diffProtein}{A data frame for the differential analysis result of
proteins.}
}
\value{
A data frame.
}
\description{
\code{adjustProteinLevel} performs the adjustment with respect to protein
abundance.
}
