% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_site.R
\name{annotate_site}
\alias{annotate_site}
\title{Annotate modification site.}
\usage{
annotate_site(aa_idx, aa_idx_all, residue)
}
\arguments{
\item{aa_idx}{An integer vector. Location of the sites.}

\item{aa_idx_all}{An integer vector. Location of all sites of the protein.}

\item{residue}{A string vector. Amino acid residue.}
}
\value{
A string.
}
\description{
\code{annotate_site} annotates modified sites as their residues and locations.
}
\examples{
annotate_site(10, c(10, 100), "K")
annotate_site(c(10, 13), c(10, 13, 100), c("K", "K"))

}
