% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_converters.R
\name{.fixTerminus}
\alias{.fixTerminus}
\title{Fix terminus location adjustments}
\usage{
.fixTerminus(data, terminus_id, unmod_pep_col)
}
\arguments{
\item{data}{data.table containing peptide data}

\item{terminus_id}{character string identifying the terminus
(e.g. N-terminus)}

\item{unmod_pep_col}{character string specifying the column name containing
unmodified peptide sequences}
}
\value{
data.table with corrected Start positions
}
\description{
Fix terminus location adjustments
}
