% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_converters.R
\name{.joinFasta}
\alias{.joinFasta}
\title{Add FASTA data into dataframe}
\usage{
.joinFasta(data, fasta_file)
}
\arguments{
\item{data}{data.table}

\item{fasta_file}{string or data.table}
}
\value{
data.table
}
\description{
Add FASTA data into dataframe
}
\keyword{internal}
