% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateAbundance}
\alias{estimateAbundance}
\title{Estimate log2-abundances of PTM sites or proteins}
\usage{
estimateAbundance(df, fctBatch = FALSE, perProtein = FALSE)
}
\arguments{
\item{df}{A data frame with columns of \code{protein}, \code{site},
\code{group}, \code{run}, \code{log2inty}, and possibly, \code{batch}.}

\item{fctBatch}{A logical. \code{TRUE} considers a fixed batch effect,
\code{FALSE} otherwise. Default is \code{FALSE}.}

\item{perProtein}{A logical. \code{TRUE} ignores the site-level information
for PTM and considers protein as a whole, \code{FALSE} otherwise. Default
is \code{FALSE}.}
}
\value{
A list of two elements named \code{PTM} and \code{PROTEIN}. The
\code{PTM} list has four elements: \code{protein} (a string vector of
protein names), \code{site} (a string vector of PTM sites), \code{param}
(a list of model parameter estimates for each site), and \code{df} (a
numeric vector of degrees of freedom for each model). The \code{PROTEIN}
list includes all as in \code{PTM}, except \code{site}.
}
\description{
\code{estimateAbundance} takes as input the summarized log2-intensities for
each PTM site, performs statistical modeling for the abundance of the site,
and returns the estimates of model parameters for all sites in all
experimental conditions.
}
\examples{
sim <- PTMsimulateExperiment(
    nGroup=2, nRep=2, nProtein=1, nSite=1, nFeature=5,
    list(PTM=25, PROTEIN=25), list(PTM=c(0, 1), PROTEIN=c(0, 1)),
    list(PTM=0.2, PROTEIN=0.2), list(PTM=0.05, PROTEIN=0.05)
)
s <- PTMsummarize(sim)
estimateAbundance(s[["PTM"]])
estimateAbundance(s[["PROTEIN"]], perProtein=TRUE)

}
