% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{extractMeanDiff}
\alias{extractMeanDiff}
\title{Compare mean abundances for all PTM sites (or proteins) across conditions}
\usage{
extractMeanDiff(data, controls, cases, per_protein = FALSE)
}
\arguments{
\item{data}{A list of abundance estimates with the following elements:
\code{protein}, \code{site}, \code{param}, and \code{df}. \code{site} may
be excluded when performing per-protein analysis (\code{per_protein} is
\code{TRUE}).}

\item{controls}{A string vector of control groups in the comparisons.}

\item{cases}{A string vector of case groups.}

\item{per_protein}{A logical. \code{TRUE} ignores the site-level information
for PTM and considers protein as a whole, \code{FALSE} performs site-level
analysis. Default is \code{FALSE}.}
}
\value{
A data frame.
}
\description{
\code{extractMeanDiff} performs significance analysis for detection of
changes in PTM mean abundances between conditions.
}
