% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_group.R
\name{lm_group}
\alias{lm_group}
\title{Fit a linear model with group effect (and batch effect if present).}
\usage{
lm_group(onesite, w_batch = FALSE)
}
\arguments{
\item{onesite}{A data frame containing columns \code{log2inty},
\code{group} for one modification site.}

\item{w_batch}{A logical scalar. \code{TRUE} includes batch effect,
\code{FALSE} otherwise. Default is \code{FALSE}.}
}
\value{
An object of class \code{lm}.
}
\description{
\code{lm_group} fits and returns a linear model with \code{group} effect.
}
\examples{
x <- data.frame(group = rep(c("1", "2"), 3), log2inty = rep(c(10, 12), 3) + rnorm(6))
lm_group(x)
}
