% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_mod.R
\name{locate_mod}
\alias{locate_mod}
\title{Locate modified sites.}
\usage{
locate_mod(pep_seq, pep_idx, mod_res_symbol)
}
\arguments{
\item{pep_seq}{A string. Peptide sequence.}

\item{pep_idx}{A matrix. Starting and ending indices of the peptide.}

\item{mod_res_symbol}{A string. Modification residue and denoted symbol.}
}
\value{
A string.
}
\description{
\code{locate_mod} locates modified sites on a peptide.
}
\examples{
locate_mod(pep_seq, pep_idx, mod_res_symbol)

}
