% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_mod.R
\name{locate_mod}
\alias{locate_mod}
\title{Locate modified sites with a peptide.}
\usage{
locate_mod(peptide, aa_start, residue_symbol)
}
\arguments{
\item{peptide}{A string. Peptide sequence.}

\item{aa_start}{An integer. Starting index of the peptide.}

\item{residue_symbol}{A string. Modification residue and denoted symbol.}
}
\value{
A string.
}
\description{
\code{locate_mod} locates modified sites with a peptide.
}
\examples{
locate_mod(peptide, aa_start, residue_symbol)
}
