% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_site.R
\name{locate_site}
\alias{locate_site}
\title{Locate potential modification sites with a peptide.}
\usage{
locate_site(peptide, pep_idx, residue)
}
\arguments{
\item{peptide}{A string. Peptide sequence.}

\item{pep_idx}{A matrix. Starting and ending indices of the peptide.}

\item{residue}{A string. Amino acid residue for the modification.}
}
\value{
A string.
}
\description{
\code{locate_site} locates modifiable sites with a peptide.
}
\examples{
idx <- matrix(c(2, 5), nrow = 1, dimnames = list(NULL, c("start", "end")))
locate_site("ABCA", idx, "A")
locate_site("ABCA", idx, "A|C")

}
