% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_site.R
\name{locate_site}
\alias{locate_site}
\title{Locate potential modification sites with a peptide.}
\usage{
locate_site(peptide, aa_start, residue)
}
\arguments{
\item{peptide}{A string. Peptide sequence.}

\item{aa_start}{An integer. Starting index of the peptide.}

\item{residue}{A string. Amino acid residue for the modification.}
}
\value{
A string.
}
\description{
\code{locate_site} locates modifiable sites with a peptide.
}
\examples{
locate_site("ABCA", 2, "A")
locate_site("ABCA", 2, "A|C")

}
