% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{MaxQtoMSstatsPTMFormat}
\alias{MaxQtoMSstatsPTMFormat}
\title{Convert output of TMT labeled MaxQuant experiment into MSstatsPTM format}
\usage{
MaxQtoMSstatsPTMFormat(
  sites.data,
  annotation.ptm,
  evidence = NULL,
  proteinGroups = NULL,
  annotation.prot = NULL,
  mod.num = "Single",
  TMT.keyword = "TMT",
  ptm.keyword = "phos",
  which.proteinid.ptm = "Protein",
  which.proteinid.protein = "Leading.razor.protein",
  removeMpeptides = FALSE
)
}
\arguments{
\item{sites.data}{modified peptide output from MaxQuant. For example, a
phosphorylation experiment would require the Phospho(STY)Sites.txt file}

\item{annotation.ptm}{data frame annotation file for the ptm level data.
Contains column Run, Fraction, TechRepMixture, Mixture, Channel,
BioReplicate, Condition.}

\item{evidence}{for global protein dataset. name of 'evidence.txt' data,
which includes feature-level data.}

\item{proteinGroups}{for global protein dataset, name of 'proteinGroups.txt'
data.}

\item{annotation.prot}{data frame annotation file for the protein level data.
Contains column Run, Fraction, TechRepMixture, Mixture, Channel,
BioReplicate, Condition.}

\item{mod.num}{For modified peptide dataset. The number modifications per
peptide to be used. If "Single", only peptides with one modification will be
used. Otherwise "Total" can be selected which does not cap the number of
modifications per peptide. "Single" is the default. Selecting "Total" may
confound the effect of different modifications.}

\item{TMT.keyword}{the sub-name of columns in sites.data file. Default is
\code{TMT}. This corresponds to the columns in the format
\code{Reporter.intensity.corrected.1.TMT1phos___1}. Specifically, this parameter
indicates the first section of the string \code{TMT1phos} (Before the mixture
number). If \code{TMT} is present in the string, set this value to \code{TMT}. Else if
\code{TMT} is not there (ie string is in the format \verb{1phos}) leave this parameter
as an empty string ('').}

\item{ptm.keyword}{the sub-name of columns in the sites.data file. Default is
\code{phos}. This corresponds to the columns in the format
\code{Reporter.intensity.corrected.1.TMT1phos___1}. Specifically, this parameter
indicates the second section of the string \code{TMT1phos} (After the mixture
number). If the string is present, set this parameter. Else if this part of
the string is empty (ie string is in the format \code{TMT1}) leave this parameter
as an empty string ('').}

\item{which.proteinid.ptm}{For PTM dataset, which column to use for protein
name. Use 'Proteins'(default) column for protein name. 'Leading.proteins' or
'Leading.razor.protein' or 'Gene.names' can be used instead to get the
protein ID with single protein. However, those can potentially have the
shared peptides.}

\item{which.proteinid.protein}{For Protein dataset, which column to use for
protein name. Same options as above.}

\item{removeMpeptides}{If Oxidation (M) modifications should be removed.
Default is TRUE.}
}
\value{
a list of two data.tables named 'PTM' and 'PROTEIN' in the format
required by MSstatsPTM.
}
\description{
Takes as input TMT experiments from MaxQ and converts the data into the
format needed for MSstatsPTM. Requires only the modified file from MaxQ (for
example Phospho(STY)Sites) and an annotation file for PTM data. To adjust
modified peptides for changes in global protein level, unmodified TMT
experimental data must also be returned.
}
\examples{

head(raw.input.tmt$PTM)
head(raw.input.tmt$PROTEIN)

}
