% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstatsPTM.R
\docType{package}
\name{MSstatsPTM}
\alias{MSstatsPTM}
\title{MSstatsPTM: A package for statistical characterization of PTMs}
\description{
The MSstatsPTM package provides four main functions for quantitative analysis
of PTMs
}
\details{
Quantitative analyses of PTMs are supported by four main functions of
\emph{MSstatsPTM}:
}
\section{Normalization}{

\code{PTMnormalize()} normalizes the quantified peak intensities to correct
systematic variation across MS runs.
}

\section{Summarization}{

\code{PTMsummarize()} summarizes log2-intensities of spectral features
(i.e., precursor ions in DDA, fragments in DIA, or transitions in SRM) into
one value per PTM site per run or one value per protein per run.
}

\section{Estimation}{

\code{PTMestimate()} takes as input the summarized log2-intensities for
each PTM site, performs statistical modeling for the log2-abundance of the
site, and returns the estimates of model parameters for all PTM sites in all
experimental conditions.
}

\section{Comparison}{

\code{PTMcompareMeans()} performs statistical testing for detecting changes
in PTM mean abundances between conditions.
}

