% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstatsPTM.R
\docType{package}
\name{MSstatsPTM}
\alias{MSstatsPTM}
\title{MSstatsPTM: A package for statistical characterization of PTMs.}
\description{
The MSstatsPTM package provides three categories of important functions:
summarization, estimation and comparison
}
\section{Summarization functions}{

The \code{PTMsummarize} and related functions summarize log2-intensities of
spectral features for each PTM site into one value per run.
}

\section{Estimation functions}{

The \code{PTMestimateAbundance} and related functions take as input the
summarized log2-intensities for each PTM site, performs statistical modeling
for the abundance of the site, and returns the estimates of model parameters
for all sites in all experimental conditions.
}

\section{Comparison functions}{

The \code{PTMcompareMeans} and related functions perform significance
analysis for detection of changes in PTM mean abundances between conditions.
}

