% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_converters.R
\name{MSstatsPTMSiteLocator}
\alias{MSstatsPTMSiteLocator}
\title{Locate modification site number and amino acid}
\usage{
MSstatsPTMSiteLocator(
  data,
  fasta_file = NULL,
  mod_id = "\\\\*",
  mod_id_is_numeric = FALSE,
  terminus_included = FALSE,
  terminus_id = "\\\\."
)
}
\arguments{
\item{data}{\code{data.table} of enriched experimental run. Must include
\code{ProteinName}, \code{PeptideSequence}, \code{PeptideModifiedSequence}, and (optionally)
\code{Start} columns.}

\item{fasta_file}{File path to FASTA file that matches with proteins in
\code{data}. Can be either string or \code{data.table} processed with \code{tidyFasta()}
function. Default to NULL if peptide number included in \code{data}.}

\item{mod_id}{String that indicates what amino acid was modified in
\code{PeptideSequence}.}

\item{mod_id_is_numeric}{Boolean indicating if mod id is numeric. If TRUE,
\code{mod_id} will be ignored. Default is FALSE.}

\item{terminus_included}{Boolean indicating if the \code{PeptideSequence} includes
the terminus amino acid.}

\item{terminus_id}{String that indicates what the terminus amino acid is.
Default is '.'.}
}
\value{
\code{data.table} with site location added into \code{Protein} column.
}
\description{
Locate modification site number and amino acid
}
\examples{
##TODO

}
