% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{PDtoMSstatsPTMFormat}
\alias{PDtoMSstatsPTMFormat}
\title{Convert Proteome Discoverer output into MSstatsPTM format}
\usage{
PDtoMSstatsPTMFormat(
  input,
  annotation,
  fasta_path = NULL,
  protein_input = NULL,
  protein_annotation = NULL,
  fasta_protein_name = NULL,
  mod_id = "\\\\(Phospho\\\\)",
  extract_unmod_from_ptm_data = FALSE,
  useNumProteinsColumn = FALSE,
  useUniquePeptide = TRUE,
  summaryforMultipleRows = max,
  removeFewMeasurements = TRUE,
  removeOxidationMpeptides = FALSE,
  removeProtein_with1Peptide = FALSE,
  which.quantification = "Precursor.Area",
  which.proteinid = "Protein.Group.Accessions",
  which.sequence = "Sequence",
  use_log_file = TRUE,
  append = FALSE,
  verbose = TRUE,
  log_file_path = NULL
)
}
\arguments{
\item{input}{PD report corresponding with enriched experimental data.}

\item{annotation}{name of 'annotation.txt' or 'annotation.csv' data which
includes Condition, BioReplicate, Run information. 'Run' will be matched with 'Spectrum.File'.}

\item{fasta_path}{A string of path to a FASTA file, used to match PTM peptides.}

\item{protein_input}{PD report corresponding with unmodified experimental data.}

\item{protein_annotation}{Same format as \code{annotation} corresponding to unmodified data.}

\item{fasta_protein_name}{todo}

\item{mod_id}{todo}

\item{extract_unmod_from_ptm_data}{todo}

\item{useNumProteinsColumn}{TRUE removes peptides which have more than 1 in
Proteins column of PD output.}

\item{useUniquePeptide}{TRUE (default) removes peptides that are assigned
for more than one proteins. We assume to use unique peptide for each protein.}

\item{summaryforMultipleRows}{max(default) or sum - when there are multiple
measurements for certain feature and certain run, use highest or sum of
multiple intensities.}

\item{removeFewMeasurements}{TRUE (default) will remove the features that
have 1 or 2 measurements across runs.}

\item{removeOxidationMpeptides}{TRUE will remove the peptides including
'oxidation (M)' in modification. FALSE is default.}

\item{removeProtein_with1Peptide}{TRUE will remove the proteins which have
only 1 peptide and charge. FALSE is default.}

\item{which.quantification}{Use 'Precursor.Area'(default) column for
quantified intensities. 'Intensity' or 'Area' can be used instead.}

\item{which.proteinid}{Use 'Protein.Accessions'(default) column for protein
name. 'Master.Protein.Accessions' can be used instead.}

\item{which.sequence}{Use 'Sequence'(default) column for peptide sequence.
'Annotated.Sequence' can be used instead.}

\item{use_log_file}{logical. If TRUE, information about data processing will
be saved to a file.}

\item{append}{logical. If TRUE, information about data processing will be
added to an existing log file.}

\item{verbose}{logical. If TRUE, information about data processing will be
printed to the console}

\item{log_file_path}{character. Path to a file to which information about
data processing will be saved. If not provided, such a file will be created
automatically. If 'append = TRUE', has to be a valid path to a file.}
}
\value{
\code{list} of \code{data.table}
}
\description{
Import Proteome Discoverer files, identify modification site location.
}
\examples{
# TODO: add examples
}
