% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{PTMcompareMeans}
\alias{PTMcompareMeans}
\title{Compare mean abundances for all PTM sites across conditions}
\usage{
PTMcompareMeans(data, controls, cases, adjProtein = FALSE)
}
\arguments{
\item{data}{A list of two elements named \code{PTM} and \code{PROTEIN}. The
\code{PTM} list has four elements: \code{protein} (a string vector of
protein names), \code{site} (a string vector of PTM sites), \code{param}
(a list of model parameter estimates for each site), and \code{df} (a
numeric vector of degrees of freedom for each model). The \code{PROTEIN}
list includes all as in \code{PTM}, except the element \code{site}.}

\item{controls}{A string vector of control groups in the comparisons.}

\item{cases}{A string vector of case groups.}

\item{adjProtein}{A logical. \code{TRUE} performs protein-level adjustment,
\code{FALSE} otherwise. Default is \code{FALSE}.}
}
\value{
A data frame.
}
\description{
\code{PTMcompareMeans} performs significance analysis for detection of
changes in PTM mean abundances between conditions.
}
\examples{
sim <- PTMsimulateExperiment(
    nGroup=2, nRep=2, nProtein=1, nSite=1, nFeature=5,
    list(PTM=25, PROTEIN=25), list(PTM=c(0, 1), PROTEIN=c(0, 1)),
    list(PTM=0.2, PROTEIN=0.2), list(PTM=0.05, PROTEIN=0.05)
)
summarized <- PTMsummarize(sim)
estimates <- PTMestimate(summarized)
PTMcompareMeans(estimates, controls="G_1", cases="G_2")

}
