% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTMcompare.R
\name{PTMcompareMeans}
\alias{PTMcompareMeans}
\title{Compare mean abundances for all PTM sites across conditions.}
\usage{
PTMcompareMeans(data, controls, cases, adjProtein = FALSE)

PTMcompareMeans(data, controls, cases, adjProtein = FALSE)
}
\arguments{
\item{data}{A list that contain estimates for PTM site abundance
(\code{data$PTM}), and possibly, protein abundance (\code{data$Protein}).}

\item{controls}{A string vector of control groups as reference.}

\item{cases}{A string vector of case groups.}

\item{adjProtein}{A logical. \code{TRUE} performs protein-level adjustment,
\code{FALSE} otherwise. Default is \code{FALSE}.}
}
\value{
A data frame.

A data frame.
}
\description{
\code{PTMcompareMeans} performs significance analysis for detection of
changes in PTM mean abundances between conditions.

\code{PTMcompareMeans} performs significance analysis for detection of
changes in PTM mean abundances between conditions.
}
\examples{
PTMcompareMeans(df_mod, controls, cases)
PTMcompareMeans(df_mod, controls, cases)
}
