% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTMestimate.R
\name{PTMestimate}
\alias{PTMestimate}
\title{Estimate log2-abundances of PTM sites and proteins.}
\usage{
PTMestimate(data, fac_batch = FALSE)
}
\arguments{
\item{data}{A list of two data frames named \code{PTM} and \code{Protein};
each contains columns of \code{protein}, \code{site}, \code{group},
\code{run}, \code{log2inty}, and possibly, \code{batch}.}

\item{fac_batch}{A logical. \code{TRUE} considers a fixed batch effect,
\code{FALSE} otherwise. Default is \code{FALSE}.}
}
\value{
A list of two elements named \code{PTM} and \code{Protein}; each is
a list with four elements: \code{protein}, \code{site}, \code{df}, and
\code{param}.
}
\description{
\code{PTMestimate} takes as input the summarized log2-intensities
for each PTM site, performs statistical modeling for the abundance of the
site, and returns the estimates of model parameters for all sites in all
experimental conditions.
}
