% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{PTMestimate}
\alias{PTMestimate}
\title{Estimate log2-abundances of PTM sites and proteins}
\usage{
PTMestimate(data, fac_batch = FALSE)
}
\arguments{
\item{data}{A list of two data frames named \code{PTM} and \code{PROTEIN}.
The \code{PTM} data frame includes columns of \code{protein}, \code{site},
\code{group}, \code{run}, \code{log2inty}, and possibly, \code{batch}. The
\code{PROTEIN} data frame includes all columns as in \code{PTM} except
\code{site}.}

\item{fac_batch}{A logical defining the handling of batch effect for all data
or two logicals for the \code{PTM} and \code{PROTEIN} (if provided) data
separately. \code{TRUE} considers a fixed batch effect, \code{FALSE}
otherwise. Default is \code{FALSE}.}
}
\value{
A list of two lists named \code{PTM} and \code{PROTEIN}. The
\code{PTM} list has four elements: \code{protein} (a string vector of
protein names), \code{site} (a string vector of PTM sites), \code{param}
(a list of model parameter estimates for each site), and \code{df} (a
numeric vector of degrees of freedom for each model). The \code{PROTEIN}
list includes all as in \code{PTM}, except \code{site}.
}
\description{
\code{PTMestimate} takes as input the summarized log2-intensities for each
PTM site, performs statistical modeling for the abundance of the site, and
returns the estimates of model parameters for all sites in all experimental
conditions. If protein log2-intensities are availble, the same estimation
procedure is applied to each protein as well.
}
\examples{
sim <- PTMsimulateExperiment(nGroup = 2, nRep = 2, nProtein = 1, nSite = 1, nFeature = 5,
list(PTM = 25, PROTEIN = 25), list(PTM = c(0, 1), PROTEIN = c(0, 1)),
list(PTM = 0.2, PROTEIN = 0.2), list(PTM = 0.05, PROTEIN = 0.05))
s <- PTMsummarize(sim)
PTMestimate(s)

}
