% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{PTMsimulateExperiment}
\alias{PTMsimulateExperiment}
\title{Simulate PTM quantification experiments}
\usage{
PTMsimulateExperiment(
  nGroup,
  nRep,
  nProtein,
  nSite,
  nFeature,
  mu,
  delta,
  sRep,
  sPeak
)
}
\arguments{
\item{nGroup}{An integer to specify the number of conditions.}

\item{nRep}{An integer to specify the number of replicates per condition.}

\item{nProtein}{An integer to specify the number of protein.}

\item{nSite}{An integer to specify the number of PTM sites per protein.}

\item{nFeature}{An integer to specify the number of features per site.}

\item{mu}{A list of two elements named \code{PTM} and \code{PROTEIN}. Each
is a numeric representing the overall mean log2-intensity in the PTM or the
PROTEIN data.}

\item{delta}{A list of two elements named \code{PTM} and \code{PROTEIN}. Each
specifies the deviation of the mean log2-abundance of each group from the
overall mean in the PTM or the PROTEIN data.}

\item{sRep}{A list of two elements named \code{PTM} and \code{PROTEIN}. Each
is a numeric representing the standard deviation for run-to-run variation.}

\item{sPeak}{A list of two elements named \code{PTM} and \code{PROTEIN}. Each
is a numeric representing the standard deviation in peak log2-intensities.}
}
\value{
A tibble with columns of \code{protein}, \code{site}, \code{group},
\code{run}, \code{feature}, \code{log2inty}.
}
\description{
\code{PTMsimulateExperiment} simulates a PTM quantification experiment with
a list of log2-intensities of multiple spectral features, PTM sites and the
corresponding proteins, in multiple MS runs and conditions.
}
\examples{
PTMsimulateExperiment(
    nGroup=2, nRep=2, nProtein=1, nSite=1, nFeature=5,
    list(PTM=25, PROTEIN=25), list(PTM=c(0, 1), PROTEIN=c(0, 1)),
    list(PTM=0.2, PROTEIN=0.2), list(PTM=0.05, PROTEIN=0.05)
)

}
