% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectronauttoMSstatsPTMFormat.R
\name{SpectronauttoMSstatsPTMFormat}
\alias{SpectronauttoMSstatsPTMFormat}
\title{Converts raw PTM MS data from Spectronautt into the format needed for
MSstatsPTM.}
\usage{
SpectronauttoMSstatsPTMFormat(
  PTM.data,
  fasta,
  Protein.data = NULL,
  annotation = NULL,
  intensity = "PeakArea",
  filter_with_Qvalue = TRUE,
  qvalue_cutoff = 0.01,
  useUniquePeptide = TRUE,
  removeFewMeasurements = TRUE,
  removeProtein_with1Feature = FALSE,
  removeNonUniqueProteins = TRUE,
  modificationLabel = "Phospho",
  removeiRT = TRUE,
  summaryforMultipleRows = max,
  which.Conditions = "all"
)
}
\arguments{
\item{PTM.data}{name of PTM Spectronaut output, which is long-format.}

\item{fasta}{A string of path to a FASTA file, used to match PTM peptides.}

\item{Protein.data}{name of Global Protein Spectronaut output, which is long-format.}

\item{annotation}{name of 'annotation.txt' data which includes Condition,
BioReplicate, Run. If annotation is already complete in Spectronaut, use
annotation=NULL (default). It will use the annotation information from input.}

\item{intensity}{'PeakArea'(default) uses not normalized peak area.
'NormalizedPeakArea' uses peak area normalized by Spectronaut}

\item{filter_with_Qvalue}{TRUE(default) will filter out the intensities that
have greater than qvalue_cutoff in EG.Qvalue column. Those intensities will
be replaced with zero and will be considered as censored missing values for
imputation purpose.}

\item{qvalue_cutoff}{Cutoff for EG.Qvalue. default is 0.01.}

\item{useUniquePeptide}{TRUE(default) removes peptides that are assigned for
more than one proteins. We assume to use unique peptide for each protein.}

\item{removeFewMeasurements}{TRUE (default) will remove the features that
have 1 or 2 measurements across runs.}

\item{removeProtein_with1Feature}{TRUE will remove the proteins which have
only 1 feature, which is the combination of peptide, precursor charge,
fragment and charge. FALSE is default.}

\item{removeNonUniqueProteins}{TRUE will remove proteins that were not
uniquely identified. IE if the protein column contains multiple proteins
seperated by ";". TRUE is default}

\item{removeiRT}{TRUE will remove proteins that contain iRT. True is default}

\item{summaryforMultipleRows}{max(default) or sum - when there are multiple
measurements for certain feature and certain run, use highest or sum of
multiple intensities.}

\item{which.Conditions}{list of conditions to format into MSstatsPTM format.
If "all" all conditions will be used. Default is "all".}

\item{removeModifications}{TRUE will remove peptide that contain a
modification. Modification must be indicated by "[". TRUE is default}
}
\description{
Takes as as input both raw PTM and global protein outputs from Spectronaut.
}
\examples{
# Output datasets of Spectronaut
## TODO: Add examples

}
