% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{tidy_fasta}
\alias{tidy_fasta}
\title{Read and tidy a FASTA file}
\usage{
tidy_fasta(path)
}
\arguments{
\item{path}{A string of path to a FASTA file.}
}
\value{
A tibble with columns named \code{header}, \code{sequence},
\code{uniprot_ac}, \code{uniprot_iso}, \code{entry_name}.
}
\description{
\code{tidy_fasta} reads and tidys FASTA file.
}
\examples{
tidy_fasta(path)
}
