% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatePTM.R
\name{tidyFasta}
\alias{tidyFasta}
\title{Read and tidy a FASTA file}
\usage{
tidyFasta(path)
}
\arguments{
\item{path}{A string of path to a FASTA file.}
}
\value{
A data.table with columns named \code{header}, \code{sequence},
\code{uniprot_ac}, \code{uniprot_iso}, \code{entry_name}.
}
\description{
\code{tidyFasta} reads and tidys FASTA file. Use this function as the first
step in identifying modification sites.
}
\examples{
tidyFasta(system.file("extdata", "O13297.fasta", package="MSstatsPTM"))

}
